refactor(constants): split utilities/constants.py into submodule package
Replace the monolithic utilities/constants.py (1095 lines) with a
utilities/constants/ package of 13 thematic submodules. A temporary
__init__.py compat shim re-exports every public name so all existing
`from utilities.constants import X` statements continue to work
unchanged while per-team import migrations land in follow-up PRs.

Submodules: aaq, components, hco, images, instance_types, misc,
monitoring, namespaces, networking, os_matrix, storage, timeouts, virt.

Bundled cleanup:
- Drop 7 unused constants (TIMEOUT_35MIN, TIMEOUT_40SEC, TIMEOUT_90MIN,
  TIMEOUT_12HRS, NONE_STR, KUBEVIRT_VIRT_OPERATOR_UP, HOSTPATH_CSI_BASIC)
- Fix typo BREW_REGISTERY_SOURCE -> BREW_REGISTRY_SOURCE (3 files)
- Dedup VIRTCTL_CLIDOWNLOADS_KUBEVIRT_HYPERCONVERGED -> VIRTCTL_CLI_DOWNLOADS

Generated-by: Cursor Agent (Claude Sonnet 4.6)
