+++ utilities/constants/__init__.py
@@ -1,0 +1,811 @@
+# Compatibility shim: re-exports every public name from submodules so that
+# existing `from utilities.constants import X` statements continue to work
+# while the per-team import migration (PRs 2-6) is in progress.
+# This file will be emptied in PR 7 once all callers use direct submodule imports.
+#
+# NOTE: `Images` is computed at the very bottom of this file (not in images.py)
+# to avoid a circular import: the original computation called
+# get_cluster_architecture() at module level, which does a lazy
+# `from utilities.constants import KUBERNETES_ARCH_LABEL` inside
+# utilities.architecture.  With the package structure that lazy import would fail
+# if triggered while the package is still being initialised.  Placing the
+# computation after all submodule imports guarantees that KUBERNETES_ARCH_LABEL
+# (from misc.py) is already in the namespace when get_cluster_architecture() runs.
+from utilities.architecture import get_cluster_architecture as _get_cluster_architecture
+from utilities.constants.aaq import (
+    AAQ_NAMESPACE_LABEL,
+    AAQ_VIRTUAL_RESOURCES,
+    AAQ_VMI_POD_USAGE,
+    ARQ_QUOTA_HARD_SPEC,
+    LIMITS_CPU_STR,
+    LIMITS_MEMORY_STR,
+    NODE_STR,
+    POD_LIMITS_CPU,
+    POD_LIMITS_MEMORY,
+    POD_REQUESTS_CPU,
+    POD_REQUESTS_MEMORY,
+    PODS_STR,
+    QUOTA_FOR_ONE_VMI,
+    QUOTA_FOR_POD,
+    REQUESTS_CPU_STR,
+    REQUESTS_CPU_VMI_STR,
+    REQUESTS_INSTANCES_VMI_STR,
+    REQUESTS_MEMORY_STR,
+    REQUESTS_MEMORY_VMI_STR,
+    VM_CPU_CORES,
+    VM_MEMORY_GUEST,
+)
+from utilities.constants.components import (
+    AAQ_OPERATOR,
+    ALL_CNV_DAEMONSETS,
+    ALL_CNV_DEPLOYMENTS,
+    ALL_CNV_PODS,
+    ALL_HCO_RELATED_OBJECTS,
+    BRIDGE_MARKER,
+    CDI_APISERVER,
+    CDI_DEPLOYMENT,
+    CDI_KUBEVIRT_HYPERCONVERGED,
+    CDI_OPERATOR,
+    CDI_STR,
+    CDI_UPLOADPROXY,
+    CLUSTER,
+    CLUSTER_NETWORK_ADDONS_OPERATOR,
+    CNAO_OPERATOR,
+    CNV_OPERATORS,
+    CNV_PODS_NO_HPP_CSI_HPP_POOL,
+    CNV_PROMETHEUS_RULES,
+    CONFIGMAP_STR,
+    CONSOLE_PLUGIN_STR,
+    CONSOLECLIDOWNLOAD_STR,
+    CONSOLEQUICKSTART_STR,
+    CPU_MODEL_LABEL_PREFIX,
+    CREATING_VIRTUAL_MACHINE,
+    CREATING_VIRTUAL_MACHINE_FROM_VOLUME,
+    DEPLOYMENT_STR,
+    GRAFANA_DASHBOARD_KUBEVIRT_TOP_CONSUMERS,
+    HCO_BEARER_AUTH,
+    HCO_CATALOG_SOURCE,
+    HCO_OPERATOR,
+    HCO_PART_OF_LABEL_VALUE,
+    HCO_WEBHOOK,
+    HOSTPATH_PROVISIONER,
+    HOSTPATH_PROVISIONER_CSI,
+    HOSTPATH_PROVISIONER_OPERATOR,
+    HPP_POOL,
+    HYPERCONVERGED_CLUSTER,
+    HYPERCONVERGED_CLUSTER_CLI_DOWNLOAD,
+    HYPERCONVERGED_CLUSTER_OPERATOR,
+    HYPERCONVERGED_CLUSTER_OPERATOR_METRICS,
+    IMAGESTREAM_STR,
+    KUBE_CNI_LINUX_BRIDGE_PLUGIN,
+    KUBEMACPOOL_CERT_MANAGER,
+    KUBEMACPOOL_MAC_CONTROLLER_MANAGER,
+    KUBEMACPOOL_PROMETHEUS_RULE,
+    KUBEVIRT_APISERVER_PROXY,
+    KUBEVIRT_APISERVER_PROXY_NP,
+    KUBEVIRT_CLUSTER_CRITICAL,
+    KUBEVIRT_CONSOLE_PLUGIN,
+    KUBEVIRT_CONSOLE_PLUGIN_NP,
+    KUBEVIRT_CONSOLE_PLUGIN_SERVICE,
+    KUBEVIRT_HCO_NAME,
+    KUBEVIRT_HYPERCONVERGED_OPERATOR_METRICS,
+    KUBEVIRT_HYPERCONVERGED_PROMETHEUS_RULE,
+    KUBEVIRT_IPAM_CONTROLLER_MANAGER,
+    KUBEVIRT_KUBEVIRT_HYPERCONVERGED,
+    KUBEVIRT_MIGRATION_CONTROLLER,
+    KUBEVIRT_MIGRATION_OPERATOR,
+    KUBEVIRT_OPERATOR,
+    KUBEVIRT_PLUGIN,
+    KUBEVIRT_STR,
+    KUBEVIRT_UI_CONFIG,
+    KUBEVIRT_UI_CONFIG_READER,
+    KUBEVIRT_UI_CONFIG_READER_ROLE_BINDING,
+    KUBEVIRT_UI_FEATURES,
+    KUBEVIRT_USER_SETTINGS,
+    MANAGED_BY_LABEL_VALUE_OLM,
+    MIGCONTROLLER_KUBEVIRT_HYPERCONVERGED,
+    NETWORKADDONSCONFIG_STR,
+    NETWORKPOLICY_STR,
+    NGINX_CONF,
+    NODE_TYPE_WORKER_LABEL,
+    POD_STR,
+    PRIORITYCLASS_STR,
+    PROMETHEUS_RULES_STR,
+    PROMETHEUSRULE_STR,
+    RHEL8_GUEST,
+    RHEL9_GUEST,
+    RHEL9_STR,
+    RHEL10_GUEST,
+    RHEL10_STR,
+    ROLE_STR,
+    ROLEBINDING_STR,
+    ROUTE_STR,
+    SECRET_STR,
+    SERVICE_STR,
+    SERVICEACCOUNT_STR,
+    SERVICEMONITOR_STR,
+    SSP_KUBEVIRT_HYPERCONVERGED,
+    SSP_OPERATOR,
+    SSP_STR,
+    TSC_FREQUENCY,
+    UPLOAD_BOOT_SOURCE,
+    VIRT_API,
+    VIRT_CONTROLLER,
+    VIRT_EXPORTPROXY,
+    VIRT_HANDLER,
+    VIRT_LAUNCHER,
+    VIRT_OPERATOR,
+    VIRT_PLATFORM_AUTOPILOT,
+    VIRT_SYNCHRONIZATION_CONTROLLER,
+    VIRT_TEMPLATE_VALIDATOR,
+    VIRTCTL_CLI_DOWNLOADS,
+    VIRTIO_WIN,
+    VM_CONSOLE_PROXY_CLUSTER_RESOURCES,
+    VM_CONSOLE_PROXY_NAMESPACE_RESOURCES,
+    WINDOWS_BOOTSOURCE_PIPELINE,
+)
+from utilities.constants.hco import (
+    ALL_CNV_CRDS,
+    COMMON_TEMPLATES_KEY_NAME,
+    DATA_IMPORT_CRON_ENABLE,
+    DATA_SOURCE_NAME,
+    DATA_SOURCE_NAMESPACE,
+    DEFAULT_HCO_CONDITIONS,
+    DEFAULT_KUBEVIRT_CONDITIONS,
+    DEFAULT_RESOURCE_CONDITIONS,
+    ENABLE_COMMON_BOOT_IMAGE_IMPORT,
+    EXPECTED_STATUS_CONDITIONS,
+    FEATURE_GATES,
+    HCO_DEFAULT_CPU_MODEL_KEY,
+    HCO_SUBSCRIPTION,
+    HOTFIX_STR,
+    IMAGE_CRON_STR,
+    PRODUCTION_CATALOG_SOURCE,
+    RESOURCE_REQUIREMENTS_KEY_HCO_CR,
+    SSP_CR_COMMON_TEMPLATES_LIST_KEY_NAME,
+    TLS_CUSTOM_POLICY,
+    TLS_OLD_POLICY,
+    TLS_SECURITY_PROFILE,
+    VM_CLONE_CRD,
+    VM_CRD,
+    VM_EXPORT_CRD,
+    UpgradeStreams,
+)
+from utilities.constants.images import (
+    ALPINE_VERSION,
+    CIRROS_DISK_DEMO,
+    CIRROS_QCOW2_IMG,
+    FEDORA_DISK_DEMO,
+    OS_FLAVOR_ALPINE,
+    OS_FLAVOR_CIRROS,
+    OS_FLAVOR_FEDORA,
+    OS_FLAVOR_RHEL,
+    OS_FLAVOR_WIN_CONTAINER_DISK,
+    OS_FLAVOR_WINDOWS,
+    ArchImages,
+)
+from utilities.constants.instance_types import (
+    CENTOS_STREAM9_PREFERENCE,
+    CENTOS_STREAM10_PREFERENCE,
+    EXPECTED_CLUSTER_INSTANCE_TYPE_LABELS,
+    FLAVOR_STR,
+    INSTANCE_TYPE_STR,
+    LINUX_STR,
+    OS_STR,
+    PREFERENCE_STR,
+    RHEL8_PREFERENCE,
+    RHEL9_PREFERENCE,
+    RHEL10_PREFERENCE,
+    RHEL_WITH_INSTANCETYPE_AND_PREFERENCE,
+    U1_LARGE,
+    U1_MEDIUM_STR,
+    U1_SMALL,
+    WORKLOAD_STR,
+)
+from utilities.constants.misc import (
+    AMD,
+    AMD_64,
+    ARM_64,
+    AUDIT_LOGS_PATH,
+    BACKUP_STORAGE_LOCATION,
+    BASE_EXCEPTIONS_DICT,
+    CAPACITY,
+    CNV_SUPPLEMENTAL_TEMPLATES_URL,
+    CNV_TEST_RUN_IN_PROGRESS,
+    CNV_TEST_RUN_IN_PROGRESS_NS,
+    CNV_TEST_SERVICE_ACCOUNT,
+    CNV_TESTS_CONTAINER,
+    CNV_VM_SSH_KEY_PATH,
+    COUNT_FIVE,
+    CREATE_STR,
+    DEFAULT_FEDORA_REGISTRY_URL,
+    DELETE_STR,
+    DEPENDENCY_SCOPE_SESSION,
+    EIGHT_CPU_SOCKETS,
+    FILE_NAME_FOR_BACKUP,
+    FIVE_GI_MEMORY,
+    FOUR_CPU_SOCKETS,
+    FOUR_GI_MEMORY,
+    GET_STR,
+    INTEL,
+    KERNEL_DRIVER,
+    KUBECONFIG,
+    KUBERNETES_ARCH_LABEL,
+    LINUX_AMD_64,
+    LS_COMMAND,
+    MULTIARCH,
+    NET_UTIL_CONTAINER_IMAGE,
+    NODE_HUGE_PAGES_1GI_KEY,
+    NODE_ROLE_KUBERNETES_IO,
+    OC_ADM_LOGS_COMMAND,
+    ONE_CPU_CORE,
+    ONE_CPU_THREAD,
+    OUTDATED,
+    POD_CONTAINER_SPEC,
+    POD_SECURITY_CONTEXT_SPEC,
+    POD_SECURITY_NAMESPACE_LABELS,
+    PROMETHEUS_K8S,
+    QUARANTINED,
+    REGISTRY_STR,
+    RHSM_SECRET_NAME,
+    S390X,
+    SANITY_TESTS_FAILURE,
+    SECURITY_CONTEXT,
+    SETUP_ERROR,
+    SIX_CPU_SOCKETS,
+    SIX_GI_MEMORY,
+    SUPPORTED_CPU_ARCHITECTURES,
+    SUPPORTED_MULTIARCH_OPTIONS,
+    TEN_CPU_SOCKETS,
+    TEN_GI_MEMORY,
+    TEXT_TO_TEST,
+    TWELVE_GI_MEMORY,
+    TWO_CPU_CORES,
+    TWO_CPU_SOCKETS,
+    TWO_CPU_THREADS,
+    UNPRIVILEGED_PASSWORD,
+    UNPRIVILEGED_USER,
+    UPDATE_STR,
+    USED,
+    UTILITY,
+    VALUE_STR,
+    VERSION_LABEL_KEY,
+    WILDCARD_CRON_EXPRESSION,
+    WORKER_NODE_LABEL_KEY,
+    WORKERS_TYPE,
+    X86_64,
+)
+from utilities.constants.monitoring import (
+    CRITICAL_STR,
+    FIRING_STATE,
+    INFO_STR,
+    KUBELET_READY_CONDITION,
+    KUBEVIRT_HCO_HYPERCONVERGED_CR_EXISTS,
+    KUBEVIRT_HYPERCONVERGED_OPERATOR_HEALTH_STATUS,
+    KUBEVIRT_VMI_MEMORY_ACTUAL_BALLOON_BYTES,
+    KUBEVIRT_VMI_MEMORY_DOMAIN_BYTES,
+    KUBEVIRT_VMI_MEMORY_PGMAJFAULT_TOTAL,
+    KUBEVIRT_VMI_MEMORY_PGMINFAULT_TOTAL,
+    KUBEVIRT_VMI_MEMORY_SWAP_IN_TRAFFIC_BYTES,
+    KUBEVIRT_VMI_MEMORY_SWAP_OUT_TRAFFIC_BYTES,
+    KUBEVIRT_VMI_MEMORY_UNUSED_BYTES,
+    KUBEVIRT_VMI_MEMORY_USABLE_BYTES,
+    KUBEVIRT_VMI_NETWORK_RECEIVE_BYTES_TOTAL,
+    KUBEVIRT_VMI_NETWORK_RECEIVE_PACKETS_DROPPED_TOTAL,
+    KUBEVIRT_VMI_NETWORK_TRANSMIT_BYTES_TOTAL,
+    KUBEVIRT_VMI_NETWORK_TRANSMIT_PACKETS_DROPPED_TOTAL,
+    KUBEVIRT_VMI_STORAGE_FLUSH_REQUESTS_TOTAL,
+    KUBEVIRT_VMI_STORAGE_FLUSH_TIMES_SECONDS_TOTAL,
+    KUBEVIRT_VMI_STORAGE_IOPS_READ_TOTAL,
+    KUBEVIRT_VMI_STORAGE_IOPS_WRITE_TOTAL,
+    KUBEVIRT_VMI_STORAGE_READ_TRAFFIC_BYTES_TOTAL,
+    KUBEVIRT_VMI_STORAGE_WRITE_TRAFFIC_BYTES_TOTAL,
+    KUBEVIRT_VMI_VCPU_WAIT_SECONDS_TOTAL,
+    MONITORING_METRICS,
+    NONE_STRING,
+    OPERATOR_HEALTH_IMPACT_VALUES,
+    PENDING_STR,
+    WARNING_STR,
+)
+from utilities.constants.namespaces import ADP_NAMESPACE, NamespacesNames
+from utilities.constants.networking import (
+    ACTIVE_BACKUP,
+    FLAT_OVERLAY_STR,
+    IP_FAMILY_POLICY_PREFER_DUAL_STACK,
+    IPV4_STR,
+    IPV6_STR,
+    KMP_ENABLED_LABEL,
+    KMP_VM_ASSIGNMENT_LABEL,
+    KUBEMACPOOL_MAC_RANGE_CONFIG,
+    LINUX_BRIDGE,
+    OVS_BRIDGE,
+    PORT_80,
+    PUBLIC_DNS_SERVER_IP,
+    SRIOV,
+    SSH_PORT_22,
+)
+from utilities.constants.os_matrix import (
+    CONTAINER_DISK_IMAGE_PATH_STR,
+    DATA_SOURCE_STR,
+    DV_SIZE_STR,
+    IMAGE_NAME_STR,
+    IMAGE_PATH_STR,
+    LATEST_RELEASE_STR,
+    OS_VERSION_STR,
+    TEMPLATE_LABELS_STR,
+)
+from utilities.constants.storage import (
+    ACCESS_MODE,
+    BIND_IMMEDIATE_ANNOTATION,
+    BREW_REGISTRY_SOURCE,
+    CDI_CONFIGMAPS,
+    CDI_LABEL,
+    CDI_SECRETS,
+    CDI_UPLOAD,
+    CDI_UPLOAD_TMP_PVC,
+    HOTPLUG_DISK_SCSI_BUS,
+    HOTPLUG_DISK_SERIAL,
+    HOTPLUG_DISK_VIRTIO_BUS,
+    HPP_CAPABILITIES,
+    ICSP_FILE,
+    IDMS_FILE,
+    PVC,
+    SOURCE_POD,
+    VOLUME_MODE,
+    StorageClassNames,
+)
+from utilities.constants.timeouts import (
+    TCP_TIMEOUT_30SEC,
+    TIMEOUT_1MIN,
+    TIMEOUT_1SEC,
+    TIMEOUT_2MIN,
+    TIMEOUT_3MIN,
+    TIMEOUT_4MIN,
+    TIMEOUT_5MIN,
+    TIMEOUT_5SEC,
+    TIMEOUT_6MIN,
+    TIMEOUT_8MIN,
+    TIMEOUT_9MIN,
+    TIMEOUT_10MIN,
+    TIMEOUT_10SEC,
+    TIMEOUT_11MIN,
+    TIMEOUT_12MIN,
+    TIMEOUT_15MIN,
+    TIMEOUT_15SEC,
+    TIMEOUT_20MIN,
+    TIMEOUT_20SEC,
+    TIMEOUT_25MIN,
+    TIMEOUT_30MIN,
+    TIMEOUT_30SEC,
+    TIMEOUT_40MIN,
+    TIMEOUT_50MIN,
+    TIMEOUT_60MIN,
+    TIMEOUT_75MIN,
+    TIMEOUT_90SEC,
+    TIMEOUT_180MIN,
+)
+from utilities.constants.virt import (
+    CLOUD_INIT_DISK_NAME,
+    CLOUD_INIT_NO_CLOUD,
+    DISK_SERIAL,
+    DV_DISK,
+    ES_LIVE_MIGRATE_IF_POSSIBLE,
+    ES_NONE,
+    EVICTIONSTRATEGY,
+    EXCLUDED_CPU_MODELS,
+    EXCLUDED_CPU_MODELS_S390X,
+    EXCLUDED_OLD_CPU_MODELS,
+    HYPERV_FEATURES_LABELS_DOM_XML,
+    HYPERV_FEATURES_LABELS_VM_YAML,
+    LIVE_MIGRATE,
+    MIGRATION_POLICY_VM_LABEL,
+    OS_PROC_NAME,
+    REGEDIT_PROC_NAME,
+    ROOTDISK,
+    STRESS_CPU_MEM_IO_COMMAND,
+    TEKTON_AVAILABLE_PIPELINEREF,
+    TEKTON_AVAILABLE_TASKS,
+    VIRTCTL,
+    VIRTIO,
+    WIN_2K19,
+    WIN_2K22,
+    WIN_2K25,
+    WIN_10,
+    WIN_11,
+    WINDOWS_2K22_PREFERENCE,
+    WINDOWS_11_PREFERENCE,
+    WINDOWS_CUSTOMIZE_STR,
+    WINDOWS_EFI_INSTALLER_STR,
+)
+
+__all__ = [
+    "AAQ_NAMESPACE_LABEL",
+    "AAQ_OPERATOR",
+    "AAQ_VIRTUAL_RESOURCES",
+    "AAQ_VMI_POD_USAGE",
+    "ACCESS_MODE",
+    "ACTIVE_BACKUP",
+    "ADP_NAMESPACE",
+    "ALL_CNV_CRDS",
+    "ALL_CNV_DAEMONSETS",
+    "ALL_CNV_DEPLOYMENTS",
+    "ALL_CNV_PODS",
+    "ALL_HCO_RELATED_OBJECTS",
+    "ALPINE_VERSION",
+    "AMD",
+    "AMD_64",
+    "ARM_64",
+    "ARQ_QUOTA_HARD_SPEC",
+    "AUDIT_LOGS_PATH",
+    "BACKUP_STORAGE_LOCATION",
+    "BASE_EXCEPTIONS_DICT",
+    "BIND_IMMEDIATE_ANNOTATION",
+    "BREW_REGISTRY_SOURCE",
+    "BRIDGE_MARKER",
+    "CAPACITY",
+    "CDI_APISERVER",
+    "CDI_CONFIGMAPS",
+    "CDI_DEPLOYMENT",
+    "CDI_KUBEVIRT_HYPERCONVERGED",
+    "CDI_LABEL",
+    "CDI_OPERATOR",
+    "CDI_SECRETS",
+    "CDI_STR",
+    "CDI_UPLOAD",
+    "CDI_UPLOADPROXY",
+    "CDI_UPLOAD_TMP_PVC",
+    "CENTOS_STREAM9_PREFERENCE",
+    "CENTOS_STREAM10_PREFERENCE",
+    "CIRROS_DISK_DEMO",
+    "CIRROS_QCOW2_IMG",
+    "CLOUD_INIT_DISK_NAME",
+    "CLOUD_INIT_NO_CLOUD",
+    "CLUSTER",
+    "CLUSTER_NETWORK_ADDONS_OPERATOR",
+    "CNAO_OPERATOR",
+    "CNV_OPERATORS",
+    "CNV_PODS_NO_HPP_CSI_HPP_POOL",
+    "CNV_PROMETHEUS_RULES",
+    "CNV_SUPPLEMENTAL_TEMPLATES_URL",
+    "CNV_TESTS_CONTAINER",
+    "CNV_TEST_RUN_IN_PROGRESS",
+    "CNV_TEST_RUN_IN_PROGRESS_NS",
+    "CNV_TEST_SERVICE_ACCOUNT",
+    "CNV_VM_SSH_KEY_PATH",
+    "COMMON_TEMPLATES_KEY_NAME",
+    "CONFIGMAP_STR",
+    "CONSOLECLIDOWNLOAD_STR",
+    "CONSOLEQUICKSTART_STR",
+    "CONSOLE_PLUGIN_STR",
+    "CONTAINER_DISK_IMAGE_PATH_STR",
+    "COUNT_FIVE",
+    "CPU_MODEL_LABEL_PREFIX",
+    "CREATE_STR",
+    "CREATING_VIRTUAL_MACHINE",
+    "CREATING_VIRTUAL_MACHINE_FROM_VOLUME",
+    "CRITICAL_STR",
+    "DATA_IMPORT_CRON_ENABLE",
+    "DATA_SOURCE_NAME",
+    "DATA_SOURCE_NAMESPACE",
+    "DATA_SOURCE_STR",
+    "DEFAULT_FEDORA_REGISTRY_URL",
+    "DEFAULT_HCO_CONDITIONS",
+    "DEFAULT_KUBEVIRT_CONDITIONS",
+    "DEFAULT_RESOURCE_CONDITIONS",
+    "DELETE_STR",
+    "DEPENDENCY_SCOPE_SESSION",
+    "DEPLOYMENT_STR",
+    "DISK_SERIAL",
+    "DV_DISK",
+    "DV_SIZE_STR",
+    "EIGHT_CPU_SOCKETS",
+    "ENABLE_COMMON_BOOT_IMAGE_IMPORT",
+    "ES_LIVE_MIGRATE_IF_POSSIBLE",
+    "ES_NONE",
+    "EVICTIONSTRATEGY",
+    "EXCLUDED_CPU_MODELS",
+    "EXCLUDED_CPU_MODELS_S390X",
+    "EXCLUDED_OLD_CPU_MODELS",
+    "EXPECTED_CLUSTER_INSTANCE_TYPE_LABELS",
+    "EXPECTED_STATUS_CONDITIONS",
+    "FEATURE_GATES",
+    "FEDORA_DISK_DEMO",
+    "FILE_NAME_FOR_BACKUP",
+    "FIRING_STATE",
+    "FIVE_GI_MEMORY",
+    "FLAT_OVERLAY_STR",
+    "FLAVOR_STR",
+    "FOUR_CPU_SOCKETS",
+    "FOUR_GI_MEMORY",
+    "GET_STR",
+    "GRAFANA_DASHBOARD_KUBEVIRT_TOP_CONSUMERS",
+    "HCO_BEARER_AUTH",
+    "HCO_CATALOG_SOURCE",
+    "HCO_DEFAULT_CPU_MODEL_KEY",
+    "HCO_OPERATOR",
+    "HCO_PART_OF_LABEL_VALUE",
+    "HCO_SUBSCRIPTION",
+    "HCO_WEBHOOK",
+    "HOSTPATH_PROVISIONER",
+    "HOSTPATH_PROVISIONER_CSI",
+    "HOSTPATH_PROVISIONER_OPERATOR",
+    "HOTFIX_STR",
+    "HOTPLUG_DISK_SCSI_BUS",
+    "HOTPLUG_DISK_SERIAL",
+    "HOTPLUG_DISK_VIRTIO_BUS",
+    "HPP_CAPABILITIES",
+    "HPP_POOL",
+    "HYPERCONVERGED_CLUSTER",
+    "HYPERCONVERGED_CLUSTER_CLI_DOWNLOAD",
+    "HYPERCONVERGED_CLUSTER_OPERATOR",
+    "HYPERCONVERGED_CLUSTER_OPERATOR_METRICS",
+    "HYPERV_FEATURES_LABELS_DOM_XML",
+    "HYPERV_FEATURES_LABELS_VM_YAML",
+    "ICSP_FILE",
+    "IDMS_FILE",
+    "IMAGESTREAM_STR",
+    "IMAGE_CRON_STR",
+    "IMAGE_NAME_STR",
+    "IMAGE_PATH_STR",
+    "INFO_STR",
+    "INSTANCE_TYPE_STR",
+    "INTEL",
+    "IPV4_STR",
+    "IPV6_STR",
+    "IP_FAMILY_POLICY_PREFER_DUAL_STACK",
+    "KERNEL_DRIVER",
+    "KMP_ENABLED_LABEL",
+    "KMP_VM_ASSIGNMENT_LABEL",
+    "KUBECONFIG",
+    "KUBELET_READY_CONDITION",
+    "KUBEMACPOOL_CERT_MANAGER",
+    "KUBEMACPOOL_MAC_CONTROLLER_MANAGER",
+    "KUBEMACPOOL_MAC_RANGE_CONFIG",
+    "KUBEMACPOOL_PROMETHEUS_RULE",
+    "KUBERNETES_ARCH_LABEL",
+    "KUBEVIRT_APISERVER_PROXY",
+    "KUBEVIRT_APISERVER_PROXY_NP",
+    "KUBEVIRT_CLUSTER_CRITICAL",
+    "KUBEVIRT_CONSOLE_PLUGIN",
+    "KUBEVIRT_CONSOLE_PLUGIN_NP",
+    "KUBEVIRT_CONSOLE_PLUGIN_SERVICE",
+    "KUBEVIRT_HCO_HYPERCONVERGED_CR_EXISTS",
+    "KUBEVIRT_HCO_NAME",
+    "KUBEVIRT_HYPERCONVERGED_OPERATOR_HEALTH_STATUS",
+    "KUBEVIRT_HYPERCONVERGED_OPERATOR_METRICS",
+    "KUBEVIRT_HYPERCONVERGED_PROMETHEUS_RULE",
+    "KUBEVIRT_IPAM_CONTROLLER_MANAGER",
+    "KUBEVIRT_KUBEVIRT_HYPERCONVERGED",
+    "KUBEVIRT_MIGRATION_CONTROLLER",
+    "KUBEVIRT_MIGRATION_OPERATOR",
+    "KUBEVIRT_OPERATOR",
+    "KUBEVIRT_PLUGIN",
+    "KUBEVIRT_STR",
+    "KUBEVIRT_UI_CONFIG",
+    "KUBEVIRT_UI_CONFIG_READER",
+    "KUBEVIRT_UI_CONFIG_READER_ROLE_BINDING",
+    "KUBEVIRT_UI_FEATURES",
+    "KUBEVIRT_USER_SETTINGS",
+    "KUBEVIRT_VMI_MEMORY_ACTUAL_BALLOON_BYTES",
+    "KUBEVIRT_VMI_MEMORY_DOMAIN_BYTES",
+    "KUBEVIRT_VMI_MEMORY_PGMAJFAULT_TOTAL",
+    "KUBEVIRT_VMI_MEMORY_PGMINFAULT_TOTAL",
+    "KUBEVIRT_VMI_MEMORY_SWAP_IN_TRAFFIC_BYTES",
+    "KUBEVIRT_VMI_MEMORY_SWAP_OUT_TRAFFIC_BYTES",
+    "KUBEVIRT_VMI_MEMORY_UNUSED_BYTES",
+    "KUBEVIRT_VMI_MEMORY_USABLE_BYTES",
+    "KUBEVIRT_VMI_NETWORK_RECEIVE_BYTES_TOTAL",
+    "KUBEVIRT_VMI_NETWORK_RECEIVE_PACKETS_DROPPED_TOTAL",
+    "KUBEVIRT_VMI_NETWORK_TRANSMIT_BYTES_TOTAL",
+    "KUBEVIRT_VMI_NETWORK_TRANSMIT_PACKETS_DROPPED_TOTAL",
+    "KUBEVIRT_VMI_STORAGE_FLUSH_REQUESTS_TOTAL",
+    "KUBEVIRT_VMI_STORAGE_FLUSH_TIMES_SECONDS_TOTAL",
+    "KUBEVIRT_VMI_STORAGE_IOPS_READ_TOTAL",
+    "KUBEVIRT_VMI_STORAGE_IOPS_WRITE_TOTAL",
+    "KUBEVIRT_VMI_STORAGE_READ_TRAFFIC_BYTES_TOTAL",
+    "KUBEVIRT_VMI_STORAGE_WRITE_TRAFFIC_BYTES_TOTAL",
+    "KUBEVIRT_VMI_VCPU_WAIT_SECONDS_TOTAL",
+    "KUBE_CNI_LINUX_BRIDGE_PLUGIN",
+    "LATEST_RELEASE_STR",
+    "LIMITS_CPU_STR",
+    "LIMITS_MEMORY_STR",
+    "LINUX_AMD_64",
+    "LINUX_BRIDGE",
+    "LINUX_STR",
+    "LIVE_MIGRATE",
+    "LS_COMMAND",
+    "MANAGED_BY_LABEL_VALUE_OLM",
+    "MIGCONTROLLER_KUBEVIRT_HYPERCONVERGED",
+    "MIGRATION_POLICY_VM_LABEL",
+    "MONITORING_METRICS",
+    "MULTIARCH",
+    "NETWORKADDONSCONFIG_STR",
+    "NETWORKPOLICY_STR",
+    "NET_UTIL_CONTAINER_IMAGE",
+    "NGINX_CONF",
+    "NODE_HUGE_PAGES_1GI_KEY",
+    "NODE_ROLE_KUBERNETES_IO",
+    "NODE_STR",
+    "NODE_TYPE_WORKER_LABEL",
+    "NONE_STRING",
+    "OC_ADM_LOGS_COMMAND",
+    "ONE_CPU_CORE",
+    "ONE_CPU_THREAD",
+    "OPERATOR_HEALTH_IMPACT_VALUES",
+    "OS_FLAVOR_ALPINE",
+    "OS_FLAVOR_CIRROS",
+    "OS_FLAVOR_FEDORA",
+    "OS_FLAVOR_RHEL",
+    "OS_FLAVOR_WINDOWS",
+    "OS_FLAVOR_WIN_CONTAINER_DISK",
+    "OS_PROC_NAME",
+    "OS_STR",
+    "OS_VERSION_STR",
+    "OUTDATED",
+    "OVS_BRIDGE",
+    "PENDING_STR",
+    "PODS_STR",
+    "POD_CONTAINER_SPEC",
+    "POD_LIMITS_CPU",
+    "POD_LIMITS_MEMORY",
+    "POD_REQUESTS_CPU",
+    "POD_REQUESTS_MEMORY",
+    "POD_SECURITY_CONTEXT_SPEC",
+    "POD_SECURITY_NAMESPACE_LABELS",
+    "POD_STR",
+    "PORT_80",
+    "PREFERENCE_STR",
+    "PRIORITYCLASS_STR",
+    "PRODUCTION_CATALOG_SOURCE",
+    "PROMETHEUSRULE_STR",
+    "PROMETHEUS_K8S",
+    "PROMETHEUS_RULES_STR",
+    "PUBLIC_DNS_SERVER_IP",
+    "PVC",
+    "QUARANTINED",
+    "QUOTA_FOR_ONE_VMI",
+    "QUOTA_FOR_POD",
+    "REGEDIT_PROC_NAME",
+    "REGISTRY_STR",
+    "REQUESTS_CPU_STR",
+    "REQUESTS_CPU_VMI_STR",
+    "REQUESTS_INSTANCES_VMI_STR",
+    "REQUESTS_MEMORY_STR",
+    "REQUESTS_MEMORY_VMI_STR",
+    "RESOURCE_REQUIREMENTS_KEY_HCO_CR",
+    "RHEL8_GUEST",
+    "RHEL8_PREFERENCE",
+    "RHEL9_GUEST",
+    "RHEL9_PREFERENCE",
+    "RHEL9_STR",
+    "RHEL10_GUEST",
+    "RHEL10_PREFERENCE",
+    "RHEL10_STR",
+    "RHEL_WITH_INSTANCETYPE_AND_PREFERENCE",
+    "RHSM_SECRET_NAME",
+    "ROLEBINDING_STR",
+    "ROLE_STR",
+    "ROOTDISK",
+    "ROUTE_STR",
+    "S390X",
+    "SANITY_TESTS_FAILURE",
+    "SECRET_STR",
+    "SECURITY_CONTEXT",
+    "SERVICEACCOUNT_STR",
+    "SERVICEMONITOR_STR",
+    "SERVICE_STR",
+    "SETUP_ERROR",
+    "SIX_CPU_SOCKETS",
+    "SIX_GI_MEMORY",
+    "SOURCE_POD",
+    "SRIOV",
+    "SSH_PORT_22",
+    "SSP_CR_COMMON_TEMPLATES_LIST_KEY_NAME",
+    "SSP_KUBEVIRT_HYPERCONVERGED",
+    "SSP_OPERATOR",
+    "SSP_STR",
+    "STRESS_CPU_MEM_IO_COMMAND",
+    "SUPPORTED_CPU_ARCHITECTURES",
+    "SUPPORTED_MULTIARCH_OPTIONS",
+    "TCP_TIMEOUT_30SEC",
+    "TEKTON_AVAILABLE_PIPELINEREF",
+    "TEKTON_AVAILABLE_TASKS",
+    "TEMPLATE_LABELS_STR",
+    "TEN_CPU_SOCKETS",
+    "TEN_GI_MEMORY",
+    "TEXT_TO_TEST",
+    "TIMEOUT_1MIN",
+    "TIMEOUT_1SEC",
+    "TIMEOUT_2MIN",
+    "TIMEOUT_3MIN",
+    "TIMEOUT_4MIN",
+    "TIMEOUT_5MIN",
+    "TIMEOUT_5SEC",
+    "TIMEOUT_6MIN",
+    "TIMEOUT_8MIN",
+    "TIMEOUT_9MIN",
+    "TIMEOUT_10MIN",
+    "TIMEOUT_10SEC",
+    "TIMEOUT_11MIN",
+    "TIMEOUT_12MIN",
+    "TIMEOUT_15MIN",
+    "TIMEOUT_15SEC",
+    "TIMEOUT_20MIN",
+    "TIMEOUT_20SEC",
+    "TIMEOUT_25MIN",
+    "TIMEOUT_30MIN",
+    "TIMEOUT_30SEC",
+    "TIMEOUT_40MIN",
+    "TIMEOUT_50MIN",
+    "TIMEOUT_60MIN",
+    "TIMEOUT_75MIN",
+    "TIMEOUT_90SEC",
+    "TIMEOUT_180MIN",
+    "TLS_CUSTOM_POLICY",
+    "TLS_OLD_POLICY",
+    "TLS_SECURITY_PROFILE",
+    "TSC_FREQUENCY",
+    "TWELVE_GI_MEMORY",
+    "TWO_CPU_CORES",
+    "TWO_CPU_SOCKETS",
+    "TWO_CPU_THREADS",
+    "U1_LARGE",
+    "U1_MEDIUM_STR",
+    "U1_SMALL",
+    "UNPRIVILEGED_PASSWORD",
+    "UNPRIVILEGED_USER",
+    "UPDATE_STR",
+    "UPLOAD_BOOT_SOURCE",
+    "USED",
+    "UTILITY",
+    "VALUE_STR",
+    "VERSION_LABEL_KEY",
+    "VIRTCTL",
+    "VIRTCTL_CLI_DOWNLOADS",
+    "VIRTIO",
+    "VIRTIO_WIN",
+    "VIRT_API",
+    "VIRT_CONTROLLER",
+    "VIRT_EXPORTPROXY",
+    "VIRT_HANDLER",
+    "VIRT_LAUNCHER",
+    "VIRT_OPERATOR",
+    "VIRT_PLATFORM_AUTOPILOT",
+    "VIRT_SYNCHRONIZATION_CONTROLLER",
+    "VIRT_TEMPLATE_VALIDATOR",
+    "VM_CLONE_CRD",
+    "VM_CONSOLE_PROXY_CLUSTER_RESOURCES",
+    "VM_CONSOLE_PROXY_NAMESPACE_RESOURCES",
+    "VM_CPU_CORES",
+    "VM_CRD",
+    "VM_EXPORT_CRD",
+    "VM_MEMORY_GUEST",
+    "VOLUME_MODE",
+    "WARNING_STR",
+    "WILDCARD_CRON_EXPRESSION",
+    "WINDOWS_2K22_PREFERENCE",
+    "WINDOWS_11_PREFERENCE",
+    "WINDOWS_BOOTSOURCE_PIPELINE",
+    "WINDOWS_CUSTOMIZE_STR",
+    "WINDOWS_EFI_INSTALLER_STR",
+    "WIN_2K19",
+    "WIN_2K22",
+    "WIN_2K25",
+    "WIN_10",
+    "WIN_11",
+    "WORKERS_TYPE",
+    "WORKER_NODE_LABEL_KEY",
+    "WORKLOAD_STR",
+    "X86_64",
+    "ArchImages",
+    "Images",
+    "NamespacesNames",
+    "StorageClassNames",
+    "UpgradeStreams",
+]
+
+# Computed after all submodule imports so that KUBERNETES_ARCH_LABEL (from misc)
+# is already in the utilities.constants namespace when get_cluster_architecture()
+# triggers its lazy `from utilities.constants import KUBERNETES_ARCH_LABEL`.
+# TODO: remove this when utilities modules are refactored
+Images = getattr(ArchImages, next(iter(_get_cluster_architecture())).upper())
