+++ utilities/constants/aaq.py
@@ -1,0 +1,34 @@
+AAQ_VIRTUAL_RESOURCES = "VirtualResources"
+AAQ_VMI_POD_USAGE = "VmiPodUsage"
+NODE_STR = "node"
+
+AAQ_NAMESPACE_LABEL = {"application-aware-quota/enable-gating": ""}
+VM_CPU_CORES = 2
+REQUESTS_INSTANCES_VMI_STR = "requests.instances/vmi"
+REQUESTS_CPU_VMI_STR = "requests.cpu/vmi"
+REQUESTS_MEMORY_VMI_STR = "requests.memory/vmi"
+PODS_STR = "pods"
+LIMITS_CPU_STR = "limits.cpu"
+LIMITS_MEMORY_STR = "limits.memory"
+REQUESTS_CPU_STR = "requests.cpu"
+REQUESTS_MEMORY_STR = "requests.memory"
+POD_REQUESTS_CPU = 2
+POD_REQUESTS_MEMORY = "2.5Gi"
+POD_LIMITS_CPU = POD_REQUESTS_CPU * 2
+POD_LIMITS_MEMORY = f"{float(POD_REQUESTS_MEMORY[:-2]) * 2}Gi"
+VM_MEMORY_GUEST = "2Gi"
+QUOTA_FOR_POD = {
+    PODS_STR: "1",
+    LIMITS_CPU_STR: POD_LIMITS_CPU,
+    LIMITS_MEMORY_STR: POD_LIMITS_MEMORY,
+    REQUESTS_CPU_STR: POD_REQUESTS_CPU,
+    REQUESTS_MEMORY_STR: POD_LIMITS_MEMORY,
+}
+
+QUOTA_FOR_ONE_VMI = {
+    REQUESTS_INSTANCES_VMI_STR: "1",
+    REQUESTS_CPU_VMI_STR: VM_CPU_CORES,
+    REQUESTS_MEMORY_VMI_STR: VM_MEMORY_GUEST,
+}
+
+ARQ_QUOTA_HARD_SPEC = {**QUOTA_FOR_POD, **QUOTA_FOR_ONE_VMI}
