+++ utilities/constants/timeouts.py
@@ -1,0 +1,29 @@
+TIMEOUT_1SEC = 1
+TIMEOUT_5SEC = 5
+TIMEOUT_10SEC = 10
+TIMEOUT_15SEC = 15
+TIMEOUT_20SEC = 20
+TIMEOUT_30SEC = 30
+TIMEOUT_90SEC = 90
+TIMEOUT_1MIN = 60
+TIMEOUT_2MIN = 2 * 60
+TIMEOUT_3MIN = 3 * 60
+TIMEOUT_4MIN = 4 * 60
+TIMEOUT_5MIN = 5 * 60
+TIMEOUT_6MIN = 6 * 60
+TIMEOUT_8MIN = 8 * 60
+TIMEOUT_9MIN = 9 * 60
+TIMEOUT_10MIN = 10 * 60
+TIMEOUT_11MIN = 11 * 60
+TIMEOUT_12MIN = 12 * 60
+TIMEOUT_15MIN = 15 * 60
+TIMEOUT_20MIN = 20 * 60
+TIMEOUT_25MIN = 25 * 60
+TIMEOUT_30MIN = 30 * 60
+TIMEOUT_40MIN = 40 * 60
+TIMEOUT_50MIN = 50 * 60
+TIMEOUT_60MIN = 60 * 60
+TIMEOUT_75MIN = 75 * 60
+TIMEOUT_180MIN = 180 * 60
+
+TCP_TIMEOUT_30SEC = 30.0
